% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylorug_plot.R
\name{+.phylorug_plot}
\alias{+.phylorug_plot}
\title{Add a layer to a phylorug plot}
\usage{
\method{+}{phylorug_plot}(e1, e2)
}
\arguments{
\item{e1}{A \code{phylorug_plot} object.}

\item{e2}{A \code{phylorug_layer} object from \code{\link[=rug_layer]{rug_layer()}}, \code{\link[=dot_layer]{dot_layer()}},
\code{\link[=support_labels]{support_labels()}}, or \code{\link[=legend_layer]{legend_layer()}}.}
}
\value{
The modified \code{phylorug_plot} object.
}
\description{
Add a layer to a phylorug plot
}
