% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylorug-package.R
\docType{package}
\name{phylorug-package}
\alias{phylorug}
\alias{phylorug-package}
\title{phylorug: Visualize Clade Recovery and Support Across Phylogenetic Trees}
\description{
Generates rug plot visualizations directly on tree nodes to illustrate
clade recovery and support from multiple phylogenetic and phylogenomic
analyses. See \code{vignette("phylorug")} for a detailed guide.
}
\section{Core functions}{

\describe{
\item{\code{\link{read_trees}}}{Read tree files from disk}
\item{\code{\link{translate_tips}}}{Rename specimen codes to species names}
\item{\code{\link{check_taxa}}}{Diagnose taxon consistency across trees}
\item{\code{\link{prune_to_shared}}}{Prune trees to shared taxa}
\item{\code{\link{node_presence_matrix}}}{Make presence/support matrix }
\item{\code{\link{plot_phylorug}}}{Draw the rug plot on a reference tree}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdrifathahamed/phylorug}
  \item \url{https://mdrifathahamed.github.io/phylorug/}
  \item Report bugs at \url{https://github.com/mdrifathahamed/phylorug/issues}
}

}
\author{
\strong{Maintainer}: Md Rifath Ahamed \email{rifath.ahamed@helsinki.fi} (\href{https://orcid.org/0009-0005-6158-8658}{ORCID})

Authors:
\itemize{
  \item Md Rifath Ahamed \email{rifath.ahamed@helsinki.fi} (\href{https://orcid.org/0009-0005-6158-8658}{ORCID})
  \item Sergei Tarasov \email{sergei.tarasov@helsinki.fi} (\href{https://orcid.org/0000-0002-1737-9403}{ORCID})
}

Other contributors:
\itemize{
  \item J. Salvador Arias \email{salvador.ariasbecerra@helsinki.fi} (\href{https://orcid.org/0000-0002-3717-435X}{ORCID}) [contributor]
}

}
\keyword{internal}
